refactor to clean up Makelangelo class
