+++ src/main/java/com/marginallyclever/makelangelo/MainFrame.java
@@ -10,7 +10,6 @@ import com.marginallyclever.convenience.FileAccess;
 import com.marginallyclever.donatello.AddNodePanel;
 import com.marginallyclever.donatello.Donatello;
 import com.marginallyclever.donatello.actions.undoable.NodeAddAction;
-import com.marginallyclever.makelangelo.applicationsettings.GFXPreferences;
 import com.marginallyclever.makelangelo.applicationsettings.MetricsPreferences;
 import com.marginallyclever.makelangelo.donatelloimpl.DonatelloDropTarget;
 import com.marginallyclever.makelangelo.makeart.io.LoadFilePanel;
@@ -19,17 +18,11 @@ import com.marginallyclever.makelangelo.makeart.io.SaveGCode;
 import com.marginallyclever.makelangelo.makeart.io.TurtleFactory;
 import com.marginallyclever.makelangelo.paper.Paper;
 import com.marginallyclever.makelangelo.plotter.Plotter;
-import com.marginallyclever.makelangelo.plotter.plotterrenderer.PlotterRenderer;
-import com.marginallyclever.makelangelo.plotter.plotterrenderer.PlotterRendererFactory;
 import com.marginallyclever.makelangelo.plotter.plottersettings.PlotterSettings;
 import com.marginallyclever.makelangelo.plotter.plottersettings.PlotterSettingsManager;
-import com.marginallyclever.makelangelo.preview.Camera;
 import com.marginallyclever.makelangelo.preview.PreviewDropTarget;
 import com.marginallyclever.makelangelo.preview.PreviewPanel;
 import com.marginallyclever.makelangelo.turtle.Turtle;
-import com.marginallyclever.makelangelo.turtle.turtlerenderer.MarlinSimulationVisualizer;
-import com.marginallyclever.makelangelo.turtle.turtlerenderer.TurtleRenderFacade;
-import com.marginallyclever.makelangelo.turtle.turtlerenderer.TurtleRenderFactory;
 import com.marginallyclever.makelangelo.turtle.turtlerenderer.TurtleRenderer;
 import com.marginallyclever.util.PreferencesHelper;
 import org.slf4j.Logger;
@@ -39,7 +32,8 @@ import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.dnd.DropTarget;
-import java.awt.event.*;
+import java.awt.event.WindowAdapter;
+import java.awt.event.WindowEvent;
 import java.io.File;
 import java.io.IOException;
 import java.net.HttpURLConnection;
@@ -57,60 +51,37 @@ public class MainFrame extends JFrame {
     private static final Logger logger = LoggerFactory.getLogger(MainFrame.class);
     private final List<DockingPanel> windows = new ArrayList<>();
 
+    private final PreviewPanel previewPanel = new PreviewPanel();
     private final Donatello donatello = new Donatello();
     private final AddNodePanel addNodePanel = new AddNodePanel();
     private final AboutPanel aboutPanel = new AboutPanel(MakelangeloVersion.VERSION, MakelangeloVersion.DETAILED_VERSION);
 
     private static final String PREFERENCE_SAVE_PATH = "savePath";
 
-    private final Camera camera;
-    private final PlotterSettingsManager plotterSettingsManager = new PlotterSettingsManager();
-    private final Plotter myPlotter = new Plotter();
-    private final Paper myPaper = new Paper();
     private Turtle myTurtle = new Turtle();
-    private final TurtleRenderFacade myTurtleRenderer = new TurtleRenderFacade();
-    private PlotterRenderer myPlotterRenderer;
 
     private final MainMenu mainMenuBar;
-    private final PreviewPanel previewPanel = new PreviewPanel();
-    private final MakeleangeloRangeSlider rangeSlider = new MakeleangeloRangeSlider();
 
     public MainFrame() {
         super();
 
-        myPlotter.setSettings(plotterSettingsManager.getLastSelectedProfile());
-        myPaper.loadConfig();
-
-        createContentPane();
-
-        previewPanel.addListener(myPlotter);
-        addPlotterRendererToPreviewPanel();
-
-        rangeSlider.addChangeListener(e->{
-            myTurtleRenderer.setFirst(rangeSlider.getBottom());
-            myTurtleRenderer.setLast(rangeSlider.getTop());
-        });
-
-        onPlotterSettingsUpdate(myPlotter.getSettings());
-
-        logger.debug("Starting virtual camera...");
-        camera = new Camera();
+        setLocationByPlatform(true);
+        setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
 
         createAppWindow();
-        //checkSharingPermission();
-
-        Preferences preferences = PreferencesHelper.getPreferenceNode(PreferencesHelper.MakelangeloPreferenceKey.FILE);
-        if (preferences.getBoolean("Check for updates", false)) checkForUpdate(true);
-
-        setLocationByPlatform(true);
-
         initDocking();
         createDefaultLayout();
         resetDefaultLayout();
         saveAndRestoreLayout();
 
+        //checkSharingPermission();
+
+        Preferences preferences = PreferencesHelper.getPreferenceNode(PreferencesHelper.MakelangeloPreferenceKey.FILE);
+        if (preferences.getBoolean("Check for updates", false)) checkForUpdate(true);
+
+        // must appear after docks so that the Windows menu item can be populated.
         mainMenuBar = new MainMenu(this);
-        this.setJMenuBar(mainMenuBar);
+        setJMenuBar(mainMenuBar);
 
         setupDropTarget();
 
@@ -134,40 +105,6 @@ public class MainFrame extends JFrame {
         DockingPanel panel = new DockingPanel(persistentID,tabText);
         panel.add(component);
         windows.add(panel);
-    }
-
-    private void updatePlotterRenderer() {
-        try {
-            myPlotterRenderer = PlotterRendererFactory.valueOf(myPlotter.getSettings().getString(PlotterSettings.STYLE)).getPlotterRenderer();
-        } catch (Exception e) {
-            logger.error("Failed to find plotter style {}", myPlotter.getSettings().getString(PlotterSettings.STYLE));
-            myPlotterRenderer = PlotterRendererFactory.MAKELANGELO_5.getPlotterRenderer();
-        }
-    }
-
-    public void onPlotterSettingsUpdate(PlotterSettings settings) {
-        myPlotter.setSettings(settings);
-
-        TurtleRenderer turtleRenderer = TurtleRenderFactory.MARLIN_SIM.getTurtleRenderer();
-        if(turtleRenderer instanceof MarlinSimulationVisualizer msv) {
-            msv.setSettings(settings);
-            msv.reset();
-        }
-        myTurtleRenderer.setUpColor(settings.getColor(PlotterSettings.PEN_UP_COLOR));
-        myTurtleRenderer.setPenDiameter(settings.getDouble(PlotterSettings.DIAMETER));
-        // myTurtleRenderer.setDownColor() would be meaningless, the down color is stored in each Turtle.
-
-        updatePlotterRenderer();
-
-        if(previewPanel != null) previewPanel.repaint();
-    }
-
-    private void addPlotterRendererToPreviewPanel() {
-        previewPanel.addListener((gl2)->{
-            if(myPlotterRenderer!=null) {
-                myPlotterRenderer.render(gl2, myPlotter);
-            }
-        });
     }
 
     // check if we need to ask about sharing
@@ -201,7 +138,7 @@ public class MainFrame extends JFrame {
 
     public void openFile(String filename) {
         try {
-            LoadFilePanel loader = new LoadFilePanel(myPaper,filename);
+            LoadFilePanel loader = new LoadFilePanel(previewPanel.getPaper(),filename);
             loader.addActionListener((e)-> setTurtle((Turtle)(e).getSource()));
 
             if(filename == null || filename.trim().isEmpty()) throw new InvalidParameterException("filename cannot be empty");
@@ -243,7 +180,7 @@ public class MainFrame extends JFrame {
      * Load a vector and add it to the existing {@link Turtle}.
      */
     public void importFile() {
-        JFileChooser fileChooser = TurtleFactory.getFileChooser();
+        JFileChooser fileChooser = TurtleFactory.getLoadFileChooser();
         // load the last path from preferences
         Preferences preferences = PreferencesHelper.getPreferenceNode(PreferencesHelper.MakelangeloPreferenceKey.FILE);
         String lastPath = preferences.get(OpenFileChooser.KEY_PREFERENCE_LOAD_PATH, FileAccess.getWorkingDirectory());
@@ -318,90 +255,11 @@ public class MainFrame extends JFrame {
      *   System.exit(0);
      * }
      */
-
-    private void createContentPane() {
-        logger.debug("create content pane...");
-
-        JPanel contentPane = new JPanel(new BorderLayout());
-        contentPane.setOpaque(true);
-
-        logger.debug("  create PreviewPanel...");
-        previewPanel.setCamera(camera);
-        previewPanel.addListener(myPaper);
-        previewPanel.addListener(myPlotter);
-        previewPanel.addListener(myTurtleRenderer);
-        addPlotterRendererToPreviewPanel();
-
-        contentPane.add(previewPanel, BorderLayout.CENTER);
-        contentPane.add(rangeSlider, BorderLayout.SOUTH);
-
-        JToolBar toolBar = createToolBar();
-        contentPane.add(toolBar, BorderLayout.NORTH);
-    }
-
-    private JToolBar createToolBar() {
-        var bar = new JToolBar();
-
-        var buttonZoomOut = new AbstractAction() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                camera.zoom(1);
-            }
-        };
-        buttonZoomOut.putValue(Action.SHORT_DESCRIPTION,Translator.get("MenuView.zoomOut"));
-        buttonZoomOut.putValue(Action.ACCELERATOR_KEY,KeyStroke.getKeyStroke(KeyEvent.VK_MINUS, InputEvent.CTRL_DOWN_MASK));
-        buttonZoomOut.putValue(Action.SMALL_ICON,new ImageIcon(Objects.requireNonNull(getClass().getResource("/com/marginallyclever/makelangelo/icons8-zoom-out-16.png"))));
-        bar.add(buttonZoomOut);
-
-        var buttonZoomIn = new AbstractAction() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                camera.zoom(-1);
-            }
-        };
-        buttonZoomIn.putValue(Action.SHORT_DESCRIPTION,Translator.get("MenuView.zoomIn"));
-        buttonZoomIn.putValue(Action.ACCELERATOR_KEY,KeyStroke.getKeyStroke(KeyEvent.VK_EQUALS, InputEvent.CTRL_DOWN_MASK));
-        buttonZoomIn.putValue(Action.SMALL_ICON,new ImageIcon(Objects.requireNonNull(getClass().getResource("/com/marginallyclever/makelangelo/icons8-zoom-in-16.png"))));
-        bar.add(buttonZoomIn);
-
-        var buttonZoomToFit = new AbstractAction() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                camera.zoomToFit(myPaper.getPaperWidth(),myPaper.getPaperHeight());
-            }
-        };
-        buttonZoomToFit.putValue(Action.SHORT_DESCRIPTION,Translator.get("MenuView.zoomFit"));
-        buttonZoomToFit.putValue(Action.ACCELERATOR_KEY,KeyStroke.getKeyStroke(KeyEvent.VK_0, InputEvent.CTRL_DOWN_MASK));
-        buttonZoomToFit.putValue(Action.SMALL_ICON,new ImageIcon(Objects.requireNonNull(getClass().getResource("/com/marginallyclever/makelangelo/icons8-zoom-to-fit-16.png"))));
-        bar.add(buttonZoomToFit);
-
-        Action toggleAction = new AbstractAction() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                boolean b = GFXPreferences.getShowPenUp();
-                GFXPreferences.setShowPenUp(!b);
-            }
-        };
-        var checkboxShowPenUpMoves = new JToggleButton(toggleAction);
-        toggleAction.putValue(Action.SHORT_DESCRIPTION,Translator.get("GFXPreferences.showPenUp"));
-        toggleAction.putValue(Action.ACCELERATOR_KEY,KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_M, InputEvent.CTRL_DOWN_MASK));//"ctrl M"
-        toggleAction.putValue(Action.SMALL_ICON,new ImageIcon(Objects.requireNonNull(getClass().getResource("/com/marginallyclever/makelangelo/icons8-plane-16.png"))));
-        checkboxShowPenUpMoves.setSelected(GFXPreferences.getShowPenUp());
-        GFXPreferences.addListener((e)->checkboxShowPenUpMoves.setSelected ((boolean)e.getNewValue()));
-        bar.add(checkboxShowPenUpMoves);
-
-        return bar;
-    }
-
-
     //  For thread safety this method should be invoked from the event-dispatching thread.
     private void createAppWindow() {
-        logger.debug("Creating GUI...");
-
         setMainTitle("");
-        this.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
 
-        this.addWindowListener(new WindowAdapter() {
+        addWindowListener(new WindowAdapter() {
             // when someone tries to close the app, confirm it.
             @Override
             public void windowClosing(WindowEvent e) {
@@ -411,12 +269,10 @@ public class MainFrame extends JFrame {
         });
 
         try {
-            this.setIconImage(ImageIO.read(Objects.requireNonNull(Makelangelo.class.getResource("/logo-icon.png"))));
+            setIconImage(ImageIO.read(Objects.requireNonNull(Makelangelo.class.getResource("/logo-icon.png"))));
         } catch (IOException e) {
             logger.warn("Can't load icon", e);
         }
-
-        camera.zoomToFit( Paper.DEFAULT_WIDTH, Paper.DEFAULT_HEIGHT);
 
         loadPaths();
     }
@@ -425,10 +281,7 @@ public class MainFrame extends JFrame {
         int result = JOptionPane.showConfirmDialog(this, Translator.get("ConfirmQuitQuestion"),
                 Translator.get("ConfirmQuitTitle"), JOptionPane.YES_NO_OPTION, JOptionPane.PLAIN_MESSAGE);
         if (result == JOptionPane.YES_OPTION) {
-            previewPanel.removeListener(myPlotter);
-            this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-            myPlotter.getSettings().save();
-            plotterSettingsManager.setLastSelectedProfile(myPlotter.getSettings().getUID());
+            setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
             savePaths();
 
             // Run this on another thread than the AWT event queue to
@@ -436,7 +289,7 @@ public class MainFrame extends JFrame {
             // exiting
             new Thread(()->{
                 previewPanel.stop();
-                this.dispose();
+                dispose();
             }).start();
             return true;
         }
@@ -464,14 +317,12 @@ public class MainFrame extends JFrame {
         if (!title.trim().isEmpty()) {
             finalTitle = title + " - " + finalTitle;
         }
-        this.setTitle(finalTitle);
+        setTitle(finalTitle);
     }
 
     public void setTurtle(Turtle turtle) {
         myTurtle = turtle;
-        myTurtleRenderer.setTurtle(turtle);
-        int top = turtle.history.size();
-        rangeSlider.setLimits(0,top);
+        previewPanel.setTurtle(turtle);
     }
 
     public Turtle getTurtle() {
@@ -501,7 +352,7 @@ public class MainFrame extends JFrame {
     }
 
     public void createDefaultLayout() {
-        addDockingPanel("Preview","Preview",previewPanel);
+        addDockingPanel("Preview","Preview", previewPanel);
         addDockingPanel("Donatello","Donatello",donatello);
         addDockingPanel("AddNode","Add Node",addNodePanel);
         addDockingPanel("About","About",aboutPanel);
@@ -541,36 +392,36 @@ public class MainFrame extends JFrame {
 
         SaveGCode save = new SaveGCode();
         try {
-            int head = rangeSlider.getValue();
-            int tail = rangeSlider.getUpperValue();
-            save.run(getTurtle(), getPlotter(), this, head, tail);
+            int head = previewPanel.getRangeBottom();
+            int tail = previewPanel.getRangeTop();
+            save.run(getTurtle(), previewPanel.getPlotter(), this, head, tail);
         } catch(Exception e) {
             logger.error("Error while exporting the gcode", e);
             JOptionPane.showMessageDialog(this, Translator.get("SaveError") + e.getLocalizedMessage(), Translator.get("ErrorTitle"), JOptionPane.ERROR_MESSAGE);
         }
     }
 
-    public Paper getPaper() {
-        return myPaper;
+    public Plotter getPlotter() {
+        return previewPanel.getPlotter();
     }
 
-    public Plotter getPlotter() {
-        return myPlotter;
+    public Paper getPaper() {
+        return previewPanel.getPaper();
     }
 
     public TurtleRenderer getTurtleRenderer() {
-        return myTurtleRenderer.getRenderer();
+        return previewPanel.getTurtleRenderer();
     }
 
-    public Camera getCamera() {
-        return camera;
-    }
-
-    public void setTurtleRenderer(TurtleRenderer renderer) {
-        myTurtleRenderer.setRenderer(renderer);
+    public void setTurtleRenderer(TurtleRenderer turtleRenderer) {
+        previewPanel.setTurtleRenderer(turtleRenderer);
     }
 
     public PlotterSettingsManager getPlotterSettingsManager() {
-        return plotterSettingsManager;
+        return previewPanel.getPlotterSettingsManager();
+    }
+
+    public void onPlotterSettingsUpdate(PlotterSettings lastSelectedProfile) {
+        previewPanel.onPlotterSettingsUpdate(lastSelectedProfile);
     }
 }
