+++ src/main/java/com/marginallyclever/makelangelo/makeart/imageconverter/ImageConverterIterative.java
@@ -4,6 +4,7 @@ import com.jogamp.opengl.GL2;
 import com.marginallyclever.makelangelo.Translator;
 import com.marginallyclever.makelangelo.makeart.TransformedImage;
 import com.marginallyclever.makelangelo.paper.Paper;
+import com.marginallyclever.makelangelo.preview.OpenGLPanel;
 import com.marginallyclever.makelangelo.preview.PreviewListener;
 import com.marginallyclever.donatello.select.SelectToggleButton;
 import org.slf4j.Logger;
@@ -84,7 +85,7 @@ public abstract class ImageConverterIterative extends ImageConverter implements 
     public abstract void resume();
 
     /**
-     * Callback from {@link com.marginallyclever.makelangelo.preview.PreviewPanel} that it is time to render to the WYSIWYG display.
+     * Callback from {@link OpenGLPanel} that it is time to render to the WYSIWYG display.
      * @param gl2 the render context
      */
     @Override
