+++ src/main/java/com/marginallyclever/makelangelo/preview/PreviewDropTarget.java
@@ -13,7 +13,7 @@ import java.io.File;
 import java.util.List;
 
 /**
- * Allows the user to drag and drop a file onto the {@link PreviewPanel}.
+ * Allows the user to drag and drop a file onto the {@link OpenGLPanel}.
  */
 public class PreviewDropTarget extends DropTargetAdapter {
     private static final Logger logger = LoggerFactory.getLogger(PreviewDropTarget.class);
